feat(import) - Import Feed - Trennung Type - Code
