+++ client/src/types/categories.types.ts
@@ -134,18 +134,3 @@ export const PodcastCategories : Array<category> = [
     { id: "16-04", api: "Film Reviews", name: "TV & Film > Film Reviews" },
     { id: "16-05", api: "TV Reviews", name: "TV & Film > TV Reviews" }
 ]
-
-export const findCategories = (categoriesArray : string[]) : category | null => {
-    if (categoriesArray.length > 0) {
-        if (Array.isArray(categoriesArray[0]) && categoriesArray[0].length > 0) {
-            let value : string = categoriesArray[0][0];
-            if (categoriesArray[0].length === 2) {
-                value = value + ' > ' + categoriesArray[0][1];
-            }
-            const category = PodcastCategories.find((item) => item.name === value);
-            if (category)
-                return category;
-        }
-    }
-    return null;
-}
